remove tests duplicates from S.M.T.
