+++ src/System.Memory/tests/ReadOnlySpan/ToLower.cs
@@ -108,32 +108,7 @@ namespace System.SpanTests
                 Assert.Equal(expectedSource, source.ToArray());
             }
         }
-
-        [Fact]
-        public static void ToLower()
-        {
-            var expectedSource = new char[3] { 'a', 'B', 'c' };
-            var expectedDestination = new char[3] { 'a', 'b', 'c' };
-
-            {
-                ReadOnlySpan<char> source = new char[3] { 'a', 'B', 'c' };
-                Span<char> destination = new char[3] { 'x', 'Y', 'z' };
-
-                Assert.Equal(source.Length, source.ToLower(destination, CultureInfo.CurrentCulture));
-                Assert.Equal(expectedDestination, destination.ToArray());
-                Assert.Equal(expectedSource, source.ToArray());
-            }
-
-            {
-                ReadOnlySpan<char> source = new char[3] { 'a', 'B', 'c' };
-                Span<char> destination = new char[3] { 'x', 'Y', 'z' };
-
-                Assert.Equal(source.Length, source.ToLowerInvariant(destination));
-                Assert.Equal(expectedDestination, destination.ToArray());
-                Assert.Equal(expectedSource, source.ToArray());
-            }
-        }
-
+      
         [Fact]
         public static void MakeSureNoToLowerChecksGoOutOfRange()
         {
@@ -195,19 +170,7 @@ namespace System.SpanTests
                 Assert.False(true, "Wrong exception thrown: Expected " + typeof(ArgumentNullException).GetType() + ": Actual: " + wrongException.GetType());
             }
         }
-
-        [Theory]
-        [InlineData("HELLO", "hello")]
-        [InlineData("hello", "hello")]
-        [InlineData("", "")]
-        public static void ToLower(string s, string expected)
-        {
-            ReadOnlySpan<char> source = s.AsSpan();
-            Span<char> destination = new char[source.Length];
-            Assert.Equal(source.Length, source.ToLower(destination, CultureInfo.CurrentCulture));
-            Assert.Equal(expected, destination.ToString());
-        }
-
+       
         private static IEnumerable<object[]> ToLower_Culture_TestData()
         {
             yield return new object[] { "H\u0049 World", "h\u0131 world", new CultureInfo("tr-TR") };
@@ -232,79 +195,6 @@ namespace System.SpanTests
             Assert.Equal(source.Length, source.ToLower(destination, culture));
             Assert.Equal(expected, destination.ToString());
         }
-
-        [Theory]
-        [InlineData("HELLO", "hello")]
-        [InlineData("hello", "hello")]
-        [InlineData("", "")]
-        public static void ToLowerInvariant(string s, string expected)
-        {
-            ReadOnlySpan<char> source = s.AsSpan();
-            Span<char> destination = new char[source.Length];
-            Assert.Equal(source.Length, source.ToLowerInvariant(destination));
-            Assert.Equal(expected, destination.ToString());
-        }
-
-        [Fact]
-        public static void ToLowerToUpperInvariant_ASCII()
-        {
-            var asciiChars = new char[128];
-            var asciiCharsUpper = new char[128];
-            var asciiCharsLower = new char[128];
-
-            for (int i = 0; i < asciiChars.Length; i++)
-            {
-                char c = (char)i;
-                asciiChars[i] = c;
-
-                // Purposefully avoiding char.ToUpper/ToLower here so as not to use the same thing we're testing.
-                asciiCharsLower[i] = (c >= 'A' && c <= 'Z') ? (char)(c - 'A' + 'a') : c;
-                asciiCharsUpper[i] = (c >= 'a' && c <= 'z') ? (char)(c - 'a' + 'A') : c;
-            }
-
-            ReadOnlySpan<char> source = asciiChars;
-            var ascii = new string(asciiChars);
-            Span<char> destinationLower = new char[source.Length];
-            Span<char> destinationUpper = new char[source.Length];
-
-            Assert.Equal(source.Length, source.ToLowerInvariant(destinationLower));
-            Assert.Equal(source.Length, source.ToUpperInvariant(destinationUpper));
-
-            Assert.Equal(ascii.ToLowerInvariant(), destinationLower.ToString());
-            Assert.Equal(ascii.ToUpperInvariant(), destinationUpper.ToString());
-
-            Assert.Equal(ascii, source.ToString());
-        }
-
-        public static IEnumerable<object[]> UpperLowerCasing_TestData()
-        {
-            //lower, upper, Culture
-            yield return new object[] { "abcd", "ABCD", "en-US" };
-            yield return new object[] { "latin i", "LATIN I", "en-US" };
-            yield return new object[] { "turky \u0131", "TURKY I", "tr-TR" };
-            yield return new object[] { "turky i", "TURKY \u0130", "tr-TR" };
-            yield return new object[] { "\ud801\udc29", PlatformDetection.IsWindows7 ? "\ud801\udc29" : "\ud801\udc01", "en-US" };
-        }
-
-        [Theory]
-        [MemberData(nameof(UpperLowerCasing_TestData))]
-        public static void CasingTest(string lowerForm, string upperForm, string cultureName)
-        {
-            CultureInfo ci = CultureInfo.GetCultureInfo(cultureName);
-
-            ReadOnlySpan<char> sourceLower = lowerForm.AsSpan();
-            ReadOnlySpan<char> sourceUpper = upperForm.AsSpan();
-            Span<char> destinationLower = new char[sourceUpper.Length];
-            Span<char> destinationUpper = new char[sourceLower.Length];
-
-            Assert.Equal(sourceUpper.Length, sourceUpper.ToLower(destinationLower, ci));
-            Assert.Equal(sourceLower.Length, sourceLower.ToUpper(destinationUpper, ci));
-
-            Assert.Equal(upperForm.ToLower(ci), destinationLower.ToString());
-            Assert.Equal(lowerForm.ToUpper(ci), destinationUpper.ToString());
-
-            Assert.Equal(lowerForm, sourceLower.ToString());
-            Assert.Equal(upperForm, sourceUpper.ToString());
-        }
+                            
     }
 }
