+++ src/System.Memory/tests/System.Memory.Tests.csproj
@@ -118,7 +118,6 @@
     <Compile Include="ReadOnlySpan\ImplicitConversion.cs" />
     <Compile Include="ReadOnlySpan\IndexOf.byte.cs" />
     <Compile Include="ReadOnlySpan\IndexOf.char.cs" />
-    <Compile Include="ReadOnlySpan\IndexOf.charSpan.cs" />
     <Compile Include="ReadOnlySpan\IndexOf.T.cs" />
     <Compile Include="ReadOnlySpan\IndexOfAny.byte.cs" />
     <Compile Include="ReadOnlySpan\IndexOfAny.T.cs" />
